register as androidlib plugin
