+++ Plugins/Android/google-signin.androidlib.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 6e0bd497669714e2a93f168b4f08a38a
