+++ Plugins/Android/google-signin.androidlib/AndroidManifest.xml
