+++ Plugins/Android/google-signin.androidlib/AndroidManifest.xml.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0441bc804c6a0407d8b0e931914ad945
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
