+++ Plugins/Android/google-signin.androidlib/build.gradle
@@ -1,0 +1,30 @@
+apply plugin: 'com.android.library'
+
+dependencies {
+    compileOnly files("${project(':unityLibrary').projectDir}/libs/unity-classes.jar")
+    implementation 'androidx.credentials:credentials:1.5.0'
+    implementation 'androidx.credentials:credentials-play-services-auth:1.5.0'
+    implementation 'com.google.android.libraries.identity.googleid:googleid:1.1.1'
+    implementation 'com.google.android.gms:play-services-auth:21.2.0'
+}
+
+android {
+    namespace "com.google.googlesignin"
+    compileSdk 35
+
+    defaultConfig {
+        minSdk 24
+        targetSdk 24
+    }
+
+    lint {
+        abortOnError false
+    }
+
+    sourceSets {
+        main {
+            manifest.srcFile 'AndroidManifest.xml'
+            java.srcDirs = ['src/main/java']
+        }
+    }
+}
