+++ Plugins/Android/google-signin.androidlib/src.meta
