+++ Plugins/Android/google-signin.androidlib/src/main.meta
