+++ Plugins/Android/google-signin.androidlib/src/main/java.meta
