+++ Plugins/Android/google-signin.androidlib/src/main/java/com/google/googlesignin.meta
