+++ Plugins/Android/google-signin.androidlib/src/main/java/com/google/googlesignin/SignInActivity.java.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: f2901b650fb65496d8332c75cbc07742
